refactor(frontend): apply MVC pattern
