+++ src/app/components/Layout.tsx
@@ -2,7 +2,7 @@
 'use client'; // This directive indicates that this module should be treated as a Client Component
 
 import React from 'react';
-import { LayoutProps } from '../types';
+import { LayoutProps } from '../../types/types';
 
 const Layout: React.FC<LayoutProps> = ({ children }) => {
 	return (
