+++ src/app/views/index.ts
@@ -1,0 +1,2 @@
+export * from './LoginView';
+export * from './CharactersView';
