+++ src/controllers/index.ts
@@ -1,0 +1,2 @@
+export * from './AuthController';
+export * from './CharacterController';
